+++ libraries/FRAM_I2C/examples/FRAM_multilanguage_elements/elements_name.h
@@ -10,7 +10,7 @@
 
 char elements_name[74][14] =
 {
-  "neutronium",
+  "neutronium",  //  zero protons, one neutron.
   "Hydrogen",
   "Helium",
   "Lithium",